+++ packages/fluentui/react-northstar/src/index.ts
@@ -46,6 +46,8 @@ export * from './components/Button/ButtonGroup';
 export * from './components/Button/ButtonContent';
 
 export * from './components/Breadcrumb/Breadcrumb';
+export * from './components/Breadcrumb/BreadcrumbItem';
+export * from './components/Breadcrumb/BreadcrumbDivider';
 
 export * from './components/Chat/Chat';
 export * from './components/Chat/ChatItem';